Add libraries and Table of contents
In Part I  we determine the probability of a user to convert to the new page. 
In Part II, we used A/B Test (Hypotheses testing) to determine if the new page is better than the old page according to the users. We calculate the p-value to conclude if we should reject the Null Hypothesis.
In Part III we use a Regression model to analyse all variables in the dataset to see if one of them could influence the conversion decision of the user.
In the conclusion, we give a final advise based on our what the data tells us.